backend/requirements-python313.txt
# Requirements otimizados para Python 3.13
# n.Gabi - Chat Multi-Agente

# =============================================================================
# FRAMEWORK WEB (Atualizado para Python 3.13)
# =============================================================================
fastapi>=0.120.0
uvicorn[standard]>=0.30.0

# =============================================================================
# VALIDAÇÃO DE DADOS (Atualizado)
# =============================================================================
pydantic>=2.15.0
pydantic-settings>=2.5.0

# =============================================================================
# CACHE E COMUNICAÇÃO (Atualizado)
# =============================================================================
redis>=5.1.0
httpx>=0.27.0

# =============================================================================
# SUPABASE (Atualizado)
# =============================================================================
supabase>=2.0.0

# =============================================================================
# UTILITÁRIOS
# =============================================================================
python-multipart>=0.0.9
python-dotenv>=1.0.1

# =============================================================================
# MONITORAMENTO (Atualizado)
# =============================================================================
prometheus-client>=0.21.0
slowapi>=0.2.0

# =============================================================================
# TESTES (Atualizado)
# =============================================================================
pytest>=8.0.0
pytest-asyncio>=0.25.0

# =============================================================================
# DESENVOLVIMENTO (Opcional)
# =============================================================================
# black>=24.0.0
# isort>=5.13.0
# flake8>=7.0.0
# mypy>=1.8.0 